pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    {name = "Kieran Gray", email = "gray.kieran@protonmail.com"},
]
requires-python = "<3.13,>=3.12"
dependencies = [
    "alembic<2.0.0,>=1.13.2",
    "alembic-postgresql-enum<2.0.0,>=1.3.0",
    "bcrypt<5.0.0,>=4.2.0",
    "dishka<2.0.0,>=1.4.0",
    "fastapi<1.0.0,>=0.115.0",
    "orjson<4.0.0,>=3.10.7",
    "psycopg[binary]<4.0.0,>=3.2.3",
    "pydantic[email]<3.0.0,>=2.9.0",
    "pyjwt[crypto]<3.0.0,>=2.9.0",
    "rtoml<1.0.0,>=0.11.0",
    "sqlalchemy[mypy]==2.0.36",
    "uuid6<2025.0.0,>=2024.7.10",
    "uvicorn<1.0.0,>=0.32.0",
    "uvloop<1.0.0,>=0.21.0",
]

[dependency-groups]
test = [
    "coverage<8.0.0,>=7.6.1",
    "pytest<9.0.0,>=8.3.2",
    "pytest-asyncio<1.0.0,>=0.24.0",
]
dev = [
    "bandit<2.0.0,>=1.7.9",
    "debugpy>=1.8.9",
    "isort<6.0.0,>=5.13.2",
    "mypy<2.0.0,>=1.11.2",
    "pylint<4.0.0,>=3.3.0",
    "ruff<1.0.0,>=0.6.3",
]

[tool.uv]
default-groups = ["dev", "test"]  # TODO don't include in prod image

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.isort]
profile = "black"
line_length = 79

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py312"
exclude = ["alembic"]
line-length = 79

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B024",  # allow abstract base classes with no abstract methods or properties
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.ruff.lint.per-file-ignores]
"src/app/infrastructure/persistence/alembic/*.py" = ["F401"] # Unused imports

[tool.pytest.ini_options]
pythonpath = [
    "src",
]
testpaths = [
    "tests",
]
# works via Makefile, but not directly
asyncio_default_fixture_loop_scope = "auto"

[tool.coverage.run]
disable_warnings = ["no-data-collected"]
source = ["src"]

[tool.coverage.report]
omit = [
    "**/__init__.py",
    "**/alembic/*",
]
show_missing = true
